Moved everything into appropriate subdirectories
